+++ src/Classification/Classifier/Http/Html/Plugin/Tracking/AdCellTrackingClassifier.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Startwind\WebInsights\Classification\Classifier\Http\Html\Plugin\Tracking;
+
+use Startwind\WebInsights\Classification\Classifier\Http\Html\HtmlClassifier;
+use Startwind\WebInsights\Response\Html\HtmlDocument;
+
+class AdCellTrackingClassifier extends HtmlClassifier
+{
+    private const PREFIX = 'html:plugin:tracking:adcell';
+
+    protected function doHtmlClassification(HtmlDocument $htmlDocument): array
+    {
+        if ($htmlDocument->containsAny([
+            'Adcell.Tracking.track()',
+            'adcell.com/js/trad.js'
+        ])) {
+            return [self::PREFIX];
+        }
+        return [];
+    }
+}